+++ Sources/ZcashPaymentURI/ZIP321+Errors.swift
@@ -21,7 +21,7 @@ extension ZIP321.Errors {
         switch memoError {
         case .invalidBase64URL:
             return ZIP321.Errors.invalidBase64
-        case .memoTooLong, .memoEmpty, .notUTF8String:
+        case .memoTooLong, .notUTF8String:
             return ZIP321.Errors.memoBytesError(memoError, index == 0 ? nil : index)
         }
     }